+++ tests/test_cli.py
@@ -1,3 +1,5 @@
+import textwrap
+
 from click.testing import CliRunner
 from unittest.mock import patch
 
@@ -31,3 +33,18 @@ Please enter the master password for the database:
 Hooray, everything is safe!
 """
     assert result.output == expected_output
+
+
+@patch("hibpcli.cli.check_passwords_from_db")
+def test_keepass_subcommand_with_path_option(mock_check):
+    mock_check.return_value = [b'Entry: "test_title (test_user)"']
+    runner = CliRunner()
+    result = runner.invoke(
+        main, ["keepass", "--path", "tests/test.kdbx"], input="test"
+    )
+    expected_output = """\
+        Please enter the master password for the database: 
+        The passwords of following entries are leaked:
+        [b'Entry: "test_title (test_user)"']
+    """
+    assert result.output == textwrap.dedent(expected_output)