+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - update dependencies
+- add path option to keepass subcommand
 
 ## 0.1.0 (01.11.2019)